divide TOFTracer.jl in individual files per module
