+++ src/plotResults.jl
@@ -8,68 +8,27 @@ using TOFTracer2
 
 file = joinpath(pwd(),"ExampleFiles","TOFDATA","results","_result.hdf5")
 
-plotHighTimeRes = false # Plot every datapoint or only file averages
-plotFittedInsteadOfSummed = true # Use multi peak fitted data instead of raw
-smoothing = 1 # Average n samples, 1 for raw
-plotsymbol = ".-"
-isobarToPlot = 0
-#timedelay = Dates.Hour(0) # CLOUD12, ...
-timedelay = Dates.Hour(-1) # CLOUDX, CLOUD11
-
 backgroundSubstractionMode = 0 # 0=no correction, 1=minimum found in all data,  2=mean of range given below
 backgroundStart = DateTime(2016,10,02,19,14)
 backgroundEnd = DateTime(2016,10,02,19,20)
 
-
 massesToPlot = [
 # Examples for selecting what to plot:
+MasslistFunctions.massFromComposition(H=2,O=1)
 MasslistFunctions.massFromComposition(C=10,H=16,O=2)
 massLibrary.APINENE[1]
 ]
 
-measResult = ResultFileFunctions.loadResults(file, massesToLoad=massesToPlot, useAveragesOnly=!plotHighTimeRes, raw=!plotFittedInsteadOfSummed, massMatchTolerance=0.01)
-if isobarToPlot != 0
-  isobarResult = ResultFileFunctions.loadResults(file, massesToLoad=[isobarToPlot+0.3], massMatchTolerance=0.5, useAveragesOnly=!plotHighTimeRes, raw=!plotFittedInsteadOfSummed)
-  measResult=joinResultsMasses(measResult, isobarResult)
-end
 
-measResult.Times = measResult.Times .- timedelay
+tracesFig = PlotFunctions.plotTracesFromHDF5(file;
+				    	massesToPlot=massesToPlot,
+				   	plotHighTimeRes = false,
+				    	smoothing = 1,
+				    	backgroundSubstractionMode = backgroundSubstractionMode,
+				    	bg = (backgroundStart,backgroundEnd),
+				    	timedelay = Dates.Hour(0), # CLOUDX and CLOUD11 data have a delay of Dates.Hour(-1)
+				    	isobarToPlot = 0,
+				    	plotsymbol = ".-"
+				    	)
 
-fig=figure()
-ax = subplot(111)
 
-
-if (backgroundSubstractionMode == 0)
-  background=0
-elseif (backgroundSubstractionMode == 1)
-  background = minimum(InterpolationFunctions.averageSamples(measResult.Traces,smoothing),dims=1)
-elseif backgroundSubstractionMode == 2
-  import Statistics
-  background = Statistics.mean(measResult.Traces[(measResult.Times.>backgroundStart) .& (measResult.Times.<backgroundEnd),:],dims=1)
-end
-
-bgCorrectedTraces = measResult.Traces .- background
-
-semilogy(Dates.unix2datetime.(InterpolationFunctions.averageSamples(Dates.datetime2unix.(measResult.Times), smoothing)),InterpolationFunctions.averageSamples(bgCorrectedTraces,smoothing), plotsymbol) #linewidth=1)
-#semilogy(Dates.unix2datetime(InterpolationFunctions.averageSamples(Dates.datetime2unix(measResult.Times), smoothing)),InterpolationFunctions.averageSamples(bgCorrectedTraces,smoothing), plotsymbol) #linewidth=1)
-
-startTimeString = Dates.format(measResult.Times[1],"yyyy/mm/dd")
-endTimeString = Dates.format(measResult.Times[end],"yyyy/mm/dd")
-title("$startTimeString - $endTimeString")
-xlabel("Time [UTC]")
-ylabel("Signal [CPS]")
-
-legStrings = []
-for i = 1:length(measResult.MasslistMasses)
-  push!(legStrings,"m/z $(round(measResult.MasslistMasses[i],digits=3)) - $(MasslistFunctions.sumFormulaStringFromCompositionArray(measResult.MasslistCompositions[:,i]))")
-end
-
-box = ax.get_position()
-cols = 1
-
-majorformatter = matplotlib.dates.DateFormatter("%m/%d %H:%M")
-ax.xaxis.set_major_formatter(majorformatter)
-legend(legStrings)
-grid()
-
-tight_layout()
